Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -4,7 +4,7 @@ int main(){
     int n;
     float avgMarks = 0.0;
     float count = 0.0;
-    int sumMarks = 0;
+    float sumMarks = 0;
     scanf("%d",&n);
 
     for(int i=0;i<n;i++){
